a little more detailed example page
