+++ example/models/router-rules.ts
@@ -5,7 +5,11 @@ export const routerRules: IRouteRule[] = [
     path: "home",
   },
   { path: "auto-save" },
+  { path: "modal" },
   { path: "draft" },
   { path: "drawer" },
+  { path: "select" },
+  { path: "validation" },
+  { path: "custom" },
   { path: "", name: "home" },
 ];
