+++ package.json
@@ -25,7 +25,7 @@
   "devDependencies": {
     "@jimengio/jimo-icons": "^0.1.17",
     "@jimengio/router-code-generator": "^0.1.2",
-    "@jimengio/ruled-router": "^0.2.9",
+    "@jimengio/ruled-router": "^0.2.10",
     "@jimengio/shared-utils": "^0.1.3",
     "@types/is": "^0.0.21",
     "@types/lodash": "^4.14.123",
