+++ tsconfig.json
@@ -22,4 +22,5 @@
       }
     ]
   },
+  "exclude": ["./lib/*"]
 }
