+++ webpack/dev.js
@@ -21,6 +21,7 @@ module.exports = {
   },
   resolve: {
     extensions: [".tsx", ".ts", ".js"],
+    modules: [path.join(__dirname, "../example"), "node_modules"],
     alias: {
       "meson-form$": path.resolve(__dirname, "../src/form.tsx"),
     },
